// src/app/openapi-gen/model/todoItemsDTO.ts
/**
 * API for Todo App
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TodoItem } from './todoItem';


export interface TodoItemsDTO { 
    count?: number;
    listId?: string;
    fromDate?: string;
    toDate?: string;
    listName?: string;
    todoItemList?: Array<TodoItem>;
}